requirements.txt
Flask~=2.0.1
Jinja2~=3.0.1
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
beautifulsoup4~=4.9.3
certifi~=2021.5.30
chardet~=4.0.0
click~=8.0.1
colorama~=0.4.4
idna~=2.10
itsdangerous~=2.0.1
passlib~=1.7.4
pip~=21.1.3
pymongo~=3.11.4
python-dotenv~=0.18.0
requests~=2.25.1
setuptools~=57.1.0
soupsieve~=2.2.1
urllib3~=1.26.6
dnspython~=2.1.0
uwsgi